updateComponent moved
