+++ shelfsense/src/components/components/ComponentsEditModal.jsx
@@ -3,7 +3,8 @@ import { Modal, Box, Typography, Button, TextField, Select, MenuItem, FormContro
 import Grid from '@mui/material/Grid2';
 import useComponentsStore from "../../stores/useComponentsStore.js";
 import useProductsStore from "../../stores/useProductsStore.js";
-import {updateComponent, deleteComponent} from "../../util/services/ComponentService.jsx";
+import {deleteComponent} from "../../util/services/ComponentService.jsx";
+import {updateComponent} from "../../util/services/components/updateComponent.js";
 import useSnackbarStore from "../../stores/useSnackbarStore.js";
 import ConfirmDialog from "../confirmDialog/ConfirmDialog.jsx"
 import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
