+++ shelfsense/src/util/services/ComponentService.jsx
@@ -4,33 +4,6 @@ import {destroyStoresAndLogout} from "../destroyStoresAndLogout.js";
 
 const BASE_URL = `${import.meta.env.VITE_API_URL}/components`;
 
-export const updateComponent = async (id, updatedData) => {
-    const BASE_URL = `${import.meta.env.VITE_API_URL}/components/${id}`;
-
-    try {
-        const response = await fetch(BASE_URL, {
-            method: "PUT",
-            headers: { "Content-Type": "application/json" },
-            credentials: "include",
-            body: JSON.stringify(updatedData),
-        });
-
-        if (response.status === 401) {
-            await destroyStoresAndLogout();
-            return null;
-        }
-
-        if (!response.ok) {
-            console.error('Failed to update component:', response.status);
-            return null;
-        }
-        return await response.json();
-    } catch (error) {
-        console.error('Error updating component:', error);
-        return null;
-    }
-};
-
 export const deleteComponent = async (id) => {
     const BASE_URL = `${import.meta.env.VITE_API_URL}/components/${id}`;
